requirements.txt
fastapi
uvicorn
pymupdf
openai 
pinecone-client
python-docx 
nltk
python-multipart
python-dotenv
fastapi-cache2
cohere